Fix unsupported OS errors
